Changing the way "DIBS Order ID" is added
When viewing the order list in Zen Cart, all orders with a comment on the first order history status is marked as containing a customer comment.

Thus inserting "DIBS Order ID" into this will clutter up the display and control for the store owner, as all DIBS orders will be marked this way. Also the old code failed to insert this if the order status was set to 'default' in the module config.

With this change we rather insert the "DIBS Order ID" in a second insert to the order history, holding the DIBS info only. If the customer writes anything it's going into the first witch is always automatically made by Zen Cart for new orders.